Minor code cleanup

+++ src/include/Settings.h
@@ -8,7 +8,7 @@
 #ifndef SRC_SETTINGS_H_
 #define SRC_SETTINGS_H_
 
-const double RUMBLE = .33;
+const float RUMBLE = .33;
 //shooter module
 const double RAMPOWER = .85;
 const double RAMTIME = 2.5;
